Test many-to-many self

+++ src/main/java/com/example/genghiskhan/kingdom/KingdomRepository.java
@@ -3,6 +3,6 @@ package com.example.genghiskhan.kingdom;
 import com.example.genghiskhan.region.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KingdomRepository extends JpaRepository<Kingdom, Byte> {
+public interface KingdomRepository extends JpaRepository<Kingdom, Integer> {
 
 }
