+++ src/main/java/com/example/genghiskhan/region/RegionRepository.java
@@ -2,6 +2,6 @@ package com.example.genghiskhan.region;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegionRepository extends JpaRepository<Region, Byte> {
+public interface RegionRepository extends JpaRepository<Region, Integer> {
 
 }
